+++ lib/invoker/command_worker.rb
@@ -27,7 +27,7 @@ module Invoker
       tail_watchers = Invoker.tail_watchers[@command_label]
       color_line = "#{@command_label.color(color)} : #{line}"
       plain_line = "#{@command_label} : #{line}"
-      if Invoker.daemonize?
+      if Invoker.nocolors?
         Invoker::Logger.puts plain_line
       else
         Invoker::Logger.puts color_line